[ESI][Cosim][NFC] Refactor cosim to divorce capnp from DPI
Move generic functionality into Capnp library. This divorces it from the
DPI server, allowing other things to use it. Additionally, divorce
server from general functionality to make adding a client thread easier.
