+++ lib/Dialect/ESI/runtime/cosim/cosim_dpi_server/CMakeLists.txt
@@ -6,8 +6,6 @@
 
 add_library(EsiCosimDpiServer SHARED
   DpiEntryPoints.cpp
-  Server.cpp
-  Endpoint.cpp
 )
 set_target_properties(EsiCosimDpiServer
     PROPERTIES
